Add a key to an object

A key is a variable that is used to identify features within an object. These features can refer to spectra in a Spectra object, peptide-spectrum matches in a PSM object, or PSMs, peptides of proteins in a SummarizedExperiment or QFeatures object. They do not need to be unique.

The goal of a key is to identify matching features across data types, such as for example to match de MS spectra to the PSMs and peptides. These different data types can be stored together in an MsExperiment object, and matched through one or multiple keys.

# S4 method for SummarizedExperiment
addKEY(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

# S4 method for Spectra
addKEY(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

# S4 method for QFeatures
addKEY(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

# S4 method for PSM
addKEY(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

Arguments

object

An instance of class SummarizedExperiment, QFeatures, PSM or Spectra.

vars

character() referencing the variables to use to create the key. These are to be found in the object's rowData (for SummarizedExperiment and QFeatures objects) or spectraData (for Spectra objects) or in the object itself (for PSM objects). Throws an error if the variables aren't found. Default is NULL.

key

character(1) defining the key's name. Default is ".KEY".

force

logical(1) that defines is an existing key should be overwritten. Default is FALSE and an error is thrown if the key is already present.

sep

character(1) specifying the separator when multiple variables are pasted together to generate the key. Default is ".".