Import sage identification results

This function imports sage identification results as a PSMatch::PSM() object.

sagePSM(
  idTable,
  spectrum = "scannr",
  peptide = "peptide",
  protein = "proteins",
  decoy = "label",
  rank = "rank",
  score = "hyperscore",
  fdr = "spectrum_fdr",
  ...
)

Arguments

idTable

character(1) containing the path to the identification results, typically "results.sage.tsv", or a data.frame containing the identification results.

spectrum

character(1) variable name that defines a spectrum in the PSM data. Default are "scannr".

peptide

character(1) variable name that defines a peptide in the PSM data. Detaults are "peptide".

protein

protein character(1) variable name that defines a protein in the PSM data. Detaults are "proteins".

decoy

character(1) variable name that defines a decoy hit in the PSM data. Detaults are "label".

rank

character(1) variable name that defines the rank of the peptide spectrum match in the PSM data. Default is "rank".

score

character(1) variable name that defines the PSM score. default is "hyperscore".

fdr

character(1) variable name that defines the spectrum FDR (or any relevant reliability score that can be used for filtering, such as the PEP) in the PSM data. Default is "spectrum_fdr".

...

Additional arguments passed to read.delim().

Value

An instance of class PSM().

Author

Laurent Gatto

Examples

## Add the data to the package's cache if they
## aren't already available

if (!sagerAvailableData("id"))
  sagerAddData("id")
basename(f <- sagerIdData())
#> [1] "bb73d7e7ae9_5779264d15c02_subset_results.sage.tsv"

sagePSM(f)
#> PSM with 500 rows and 34 columns.
#> Spectra: 497 unique
#>   db: 331 target, 169 decoy
#>   ranks: 1:500 
#> Peptides: 500 unique, 0 multiple
#> Proteins: 490 
#> names(34): peptide proteins ... protein_fdr ms1_intensity