Sager Data

A set of function to install and access sager example data. Three datasets are available: a TMT11 quantitation tsv file, a identification tsv file, both generated by sage, and an mzML file that contains the raw data used as input to sage.

The data were downloaded from the ProteomeXchange project PXD016766 and processed with sage version 0.10.0 and subset to reduce their size. See inst/scripts/make-data.R for details on how these data were generated and subset.

sagerData(cache, rname)

sagerCache()

sagerQuantData()

sagerIdData()

sagerMzMLData()

sagerAddData(which = c("quant", "id", "mzml"), cache = sagerCache())

sagerRemoveData(which = c("quant", "id", "mzml"), cache = sagerCache())

sagerAvailableData(which = c("quant", "id", "mzml"), cache = sagerCache())

Arguments

cache

Object of class BiocFileCache. Default for is the package's cache returned by sagerCache().

rname

character(1) defining the resource name used to query cache.

which

One or several of "quant", "id" or "mzml" specifying what data to download.

Functions

sagerMzMLData() returns the path to the cached mzML file.

sagerQuantData() returns the path to the cached quantitation tab-separated file.

sagerIdData() returns the path to the cached identifcation tab-separated file.

sagerAvailableData() checks if the data are available in the package cache.

sagerAddData(which, cache) checks if the resources are already available in the cache. If not, data are downloaded and added to the cache.

sagerRemoveData(which, cache) removes the cached resource.

sagerCache() returns the package's cache, an object of class BiocFileCache.

The sagerMzMLData(), sagerQuantData() and sagerIdData() use the sagerData() function under the hood, that defined the resource name and the cache. A user shouldn't have to use sagerData().

The data files are downloaded and cached by sagerAddData() are available on zenodo (10.5281/zenodo.7804639).

Update locally cached data

The sager data files will occasionally be updated or new ones added. These changes are documented in the NEWS.md file.

To update your local cache, you can either delete it (it's location can be found with bfccache(sagerCache())) to remove and add the data by running sagerRemoveData() and sagerAddData().

References

Sage: proteomics searching so fast it seems like magic. https://github.com/lazear/sage.

Project PXD016766: www.ebi.ac.uk/pride/archive/projects/PXD016766.

Yu et al. 'Benchmarking the Orbitrap Tribrid Eclipse for Next Generation Multiplexed Proteomics' Anal. Chem. 2020, 92, 9, 6478–6485 Publication Date:April 6, 2020 DOI:10.1021/acs.analchem.9b05685.

sager package test data: DOI: 10.5281/zenodo.7804639.

Examples

## Are data already available
sagerAvailableData()
#> quant    id  mzml 
#>  TRUE  TRUE FALSE 

sagerAvailableData("id")
#>   id 
#> TRUE 

## Remove data from cache
sagerRemoveData("id")

sagerAvailableData("id")
#>    id 
#> FALSE 

## Add data (if not already available)
sagerAddData(c("id", "quant"))
#> Adding 'id' to the package cache.
#> 'quant' already available.

## Get the path to the data
sagerQuantData()
#> [1] "/github/home/.cache/R/sager/11ac55cc19ea_577925b6185b3_subset_quant.tsv"

sagerIdData()
#> [1] "/github/home/.cache/R/sager/12a713db51ea_5779264d15c02_subset_results.sage.tsv"

sagerMzMLData()
#> Warning: Data not found. See ?sagerData for details.
#> [1] NA

## Data are stored in the package's cache
sagerCache()
#> class: BiocFileCache
#> bfccache: /github/home/.cache/R/sager
#> bfccount: 2
#> For more information see: bfcinfo() or bfcquery()