Correction de Sheppard et variance

Calcul de la correction de Sheppard pour la variance.

sheppardCorrection(gp.data, order = 2, population = FALSE)

Arguments

gp.data

Un objet de classe grouped.data créé avec le package actuar.

order

Choix de l'ordre pour la correction de Sheppard. Ne peut être égal qu'à 2 ou 4.

population

Un booléen qui indique si le calcul est réalisé à partir d'une population (population=TRUE) ou à partir d'un échantillon pour estimer un paramètre d'une population (population=FALSE).

Value

Un vecteur contenant la valeur de la correction de Sheppard pour le calcul de la variance d'une distribution groupée.

Examples

lims <- c(40, 45, 50, 55, 60, 65, 70, 75)
counts <- c(1,2,3,4,0,0,1)
grouped.example <- actuar::grouped.data(Group = lims, Frequency = counts)
sheppardCorrection(grouped.example)
#> [1] 2.083333
sheppardCorrection(grouped.example, order=4)
#> Frequency 
#> -748.8163