Caractéristiques de champignons

Caractéristiques de 1000 champignons.

champignons

Format

A data frame with 1000 observations on the following 8 variables.

classes

a factor with levels comestible veneneux. Classes.

couleur_chapeau

a factor with levels blanc gris jaune marron rouge. Couleur du chapeau.

contusions

a factor with levels abime non. Contusions.

odeur

a factor with levels amande anis apre inodore. Odeur du champignon.

espacement_lamelle

a factor with levels ferme serre. Espacement des lamelles.

couleur_tige_dessus

a factor with levels blanc gris rose. Couleur du dessus de la tige.

habitat

a factor with levels foret pelouse prairies sentier urbain. Habitat.

Examples

data(champignons)
str(champignons)

#> 'data.frame':	1000 obs. of  8 variables:
#>  $ num_champ          : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ classes            : Factor w/ 2 levels "comestible","veneneux": 2 1 1 2 1 1 1 1 2 1 ...
#>  $ couleur_chapeau    : Factor w/ 5 levels "blanc","gris",..: 4 3 1 1 2 3 1 1 1 3 ...
#>  $ contusions         : Factor w/ 2 levels "abime","non": 1 1 1 1 2 1 1 1 1 1 ...
#>  $ odeur              : Factor w/ 4 levels "amande","anis",..: 3 1 2 3 4 1 1 2 3 1 ...
#>  $ espacement_lamelle : Factor w/ 2 levels "ferme","serre": 1 1 1 1 2 1 1 1 1 1 ...
#>  $ couleur_tige_dessus: Factor w/ 3 levels "blanc","gris",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ habitat            : Factor w/ 5 levels "foret","pelouse",..: 5 2 3 5 2 2 3 3 2 3 ...

summary(champignons)

#>    num_champ            classes    couleur_chapeau contusions      odeur    
#>  Min.   :   1.0   comestible:898   blanc :338      abime:793   amande :335  
#>  1st Qu.: 250.8   veneneux  :102   gris  : 94      non  :207   anis   :336  
#>  Median : 500.5                    jaune :333                  apre   :102  
#>  Mean   : 500.5                    marron:234                  inodore:227  
#>  3rd Qu.: 750.2                    rouge :  1                               
#>  Max.   :1000.0                                                             
#>  espacement_lamelle couleur_tige_dessus     habitat   
#>  ferme:798          blanc:983           foret   :100  
#>  serre:202          gris :  8           pelouse :462  
#>                     rose :  9           prairies:221  
#>                                         sentier : 78  
#>                                         urbain  :139  
#>